Add files via upload
we finished the merge, find_cast_member,  and read successfully. But I made several mistakes. First, I made a mistake to free all the memory for all_cast and some_cast. Hence the new all_cast will not point to any valid cast_members. Secondly, in my search char array, it points to some unused memory, but it doesn't mean that it points to a null space. Hence, to compare it with Null is wrong. Finally, in merge and find_cast_member, I use len to be the index, which is totally wrong, causing index out of bound and segmatation fault. But finally we done those three